requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataclasses-json==0.5.2
idna==2.10
marshmallow==3.7.1
marshmallow-enum==1.5.1
mypy==0.782
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
PySocks==1.7.1
python-dotenv==0.14.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
stringcase==1.2.0
toml==0.10.1
tweepy==3.9.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
